Reference

Contents

Reference

Contents

Index

Index

RationalVectorFitting.build_subA!

RationalVectorFitting.cplxpair

RationalVectorFitting.pole_identification

RationalVectorFitting.rational

RationalVectorFitting.recommended_init_poles

RationalVectorFitting.residue_identification

RationalVectorFitting.vector_fitting

RationalVectorFitting.build_subA! — Method

build_subA!(A1, s, poles)

Builds the submatrix with the 1 / (s - p), 1.0 and s coefficients. It is assumed that the poles are sorted by cplxpair.

RationalVectorFitting.cplxpair — Method

cplxpair(x)

To be used to sort an array by real, then complex conjugate pairs.

RationalVectorFitting.pole_identification — Method

pole_identification(s, f, poles, relaxed)

Stage 1 of the Vector Fitting.

RationalVectorFitting.rational — Method

rational(s, poles, residues, d, h)

Rational transfer function.

RationalVectorFitting.recommended_init_poles — Method

recommended_init_poles(s, Npairs)

Builds a vector of recommended initial poles sorted by cplxpair.

RationalVectorFitting.residue_identification — Method

residue_identification(s, f, poles)

Stage 2 of the Vector Fitting.

RationalVectorFitting.vector_fitting — Method

vector_fitting(s, f, init_poles; relaxed=true, force_stable=true, maxiter=20, tol=1e-12)

Fast Relaxed Vector Fitting of the array f with complex frequency s using a set of initial poles init_poles.

f can be a matrix of dimensions (Ns, Nc) and the fitting will be over its columns with a set of common poles.

relaxed controls the nontriviality constraint. See [2].

force_stable controls if unstable poles should be reflected to the semi-left complex plane.

maxiter is the maximum of iterations that will be done to try to achieve a convergence with desired tolerance tol.

References

[1] B. Gustavsen and A. Semlyen, "Rational approximation of frequency domain responses by vector fitting," in IEEE Transactions on Power Delivery, vol. 14, no. 3, pp. 1052-1061, July 1999, doi: 10.1109/61.772353.

[2] B. Gustavsen, "Improving the pole relocating properties of vector fitting," in IEEE Transactions on Power Delivery, vol. 21, no. 3, pp. 1587-1592, July 2006, doi: 10.1109/TPWRD.2005.860281.

[3] D. Deschrijver, M. Mrozowski, T. Dhaene and D. De Zutter, "Macromodeling of Multiport Systems Using a Fast Implementation of the Vector Fitting Method," in IEEE Microwave and Wireless Components Letters, vol. 18, no. 6, pp. 383-385, June 2008, doi: 10.1109/LMWC.2008.922585